% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trans.R
\name{apply_transform}
\alias{apply_transform}
\title{Transform arbitrary values.}
\usage{
apply_transform(data, t_fun, channel_pattern = "-A")
}
\arguments{
\item{data}{A cytoset to transform.}

\item{t_fun}{Transformation function accepting values and channel name.}

\item{channel_pattern}{Regex pattern for fluorescence channels to transform.}
}
\description{
Transform arbitrary values.
}
