% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating_designer.R
\name{build_boundary_gating_args}
\alias{build_boundary_gating_args}
\title{Build OpenCyto boundary gate arguments.}
\usage{
build_boundary_gating_args(min, max)
}
\arguments{
\item{min}{Numeric vector of lower bounds for two channels.}

\item{max}{Numeric vector of upper bounds for two channels.}
}
\value{
A character string suitable for the \code{gating_args} column of an
  OpenCyto gating template using \code{gating_method = "boundary"}.
}
\description{
Build OpenCyto boundary gate arguments.
}
