% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comp_cv}
\alias{comp_cv}
\title{Compute the coefficient of variation}
\usage{
comp_cv(x, log_t = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{log_t}{Whether to compute the coefficient of variation after log
transformation of positive values.}
}
\description{
Compute the coefficient of variation
}
