% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_data_subdir}
\alias{create_data_subdir}
\title{Create data subdirectory.}
\usage{
create_data_subdir(data_path)
}
\arguments{
\item{data_path}{The path to the original data directory.}
}
\value{
\code{TRUE} if a data directory was created and files were moved,
  \code{FALSE} if the move failed, or \code{NULL} if the data directory
  already existed.
}
\description{
This is a utility function for creating a subdirectory for the data in the
data path and moving the data files into the new directory. This way the
analysis results and data stay separated.
}
