% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressalyzr.R
\docType{package}
\name{expressalyzr-package}
\alias{expressalyzr}
\alias{expressalyzr-package}
\title{expressalyzr: Automated gating and compensation of flow cytometry data}
\description{
The expressalyzr package provides a reproducible pipeline for gating,
compensation, quality control, and event-level export of flow cytometry data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ariwor/expressalyzr}
  \item Report bugs at \url{https://github.com/Ariwor/expressalyzr/issues}
}

}
\author{
\strong{Maintainer}: Timothy Frei \email{frei.timothy@gmail.com}

Authors:
\itemize{
  \item Asterios Arampatzis \email{aarampatzis@ethz.ch}
}

}
