% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressalyzr.R
\docType{package}
\name{expressalyzr}
\alias{expressalyzr}
\title{expressalyzr: A package for automated gating and compensation of flow
cytometry data.}
\description{
The expressalyzr package provides useful functions and a general pipeline for
automated gating and compensation of flow cytometry data.
}
\section{Foo functions}{
 The foo functions ...
}

