% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{expressalyzr_palette}
\alias{expressalyzr_palette}
\title{expressalyzr plot palette.}
\usage{
expressalyzr_palette()
}
\value{
A named character vector of hex colours.
}
\description{
expressalyzr plot palette.
}
