% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_so.R
\name{filter_density}
\alias{filter_density}
\title{Filter data according to a density threshold.}
\usage{
filter_density(data, channels, bins, th)
}
\arguments{
\item{data}{A cytoset to filter.}

\item{channels}{Character vector of channels used for density filtering.}

\item{bins}{Number of bins used to estimate local density.}

\item{th}{Density threshold.}
}
\description{
Filter data according to a density threshold.
}
