% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_fcs}
\alias{load_fcs}
\title{Load .fcs files}
\usage{
load_fcs(file_path)
}
\arguments{
\item{file_path}{The path to the .fcs files that are to be loaded.}
}
\value{
A GatingSet object.
}
\description{
Load .fcs files into a cytoset.
}
\examples{
path <- system.file("extdata", "example_fcs_files", package = "expressalyzr", mustWork = TRUE)
gating_set <- load_fcs(path)
}
