% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mefl_dt}
\alias{mefl_dt}
\title{Data for molecules of equivalent fluorescin (MEFL) per peak.}
\format{
A data table with 8 rows and 2 variables:
\describe{
  \item{Peak}{number of peak, from dimmest to brightest}
  \item{MEFL}{molecules of equivalent fluorescin, in molecules}
}
}
\source{
Sheet that came with the beads.
}
\usage{
mefl_dt
}
\description{
Data for MEFL per peak for RCP-30-5A Lot #AK04.
}
\keyword{datasets}
