% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gate_hierarchy}
\alias{plot_gate_hierarchy}
\title{Plot the gating hierarchy.}
\usage{
plot_gate_hierarchy(gs)
}
\arguments{
\item{gs}{A \code{GatingSet} object after gating.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the gating hierarchy.
}
