% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_qc_overview}
\alias{plot_qc_overview}
\title{Plot a QC overview.}
\usage{
plot_qc_overview(qc_results)
}
\arguments{
\item{qc_results}{A QC list returned by \code{\link{collect_qc}}, or an
event-count table returned by \code{\link{compute_event_counts}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a QC overview.
}
