% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trans.R
\name{run_fit}
\alias{run_fit}
\title{Fit bead model to mean peak data.}
\usage{
run_fit(x, y, f)
}
\arguments{
\item{x}{Known MEFL values.}

\item{y}{Observed channel intensities.}

\item{f}{Objective function used for residuals.}
}
\description{
Fit bead model to mean peak data.
}
