% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_expressalyzr}
\alias{theme_expressalyzr}
\title{expressalyzr ggplot theme.}
\usage{
theme_expressalyzr(base_size = 11)
}
\arguments{
\item{base_size}{Base font size.}
}
\value{
A ggplot2 theme.
}
\description{
expressalyzr ggplot theme.
}
