% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_config}
\alias{load_config}
\title{Create and/or load configuration file.}
\usage{
load_config(data_path)
}
\arguments{
\item{data_path}{The path to the original data directory.}
}
\description{
A configuration file is created from a template and opened for editing.
After editing is complete or if a configuration file already exists
the file is loaded.
}
