% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the data processing pipline}
\usage{
run_pipeline()
}
\value{

}
\description{
Run the data processing pipline that links together other functions in the
 expressalyzr package.
}
\examples{
 run_pipeline()
}
