% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_so.R
\name{spillover_matrix}
\alias{spillover_matrix}
\title{Compute the spillover matrix based on single color transfection controls.}
\usage{
spillover_matrix(data, cont_ind, comp_pattern, threshold, manual_comp)
}
\description{
Compute the spillover matrix based on single color transfection controls.
}
